MOve CharMeasure into ui/
