+++ src/SelectionManager.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { assert } from 'chai';
-import { CharMeasure } from './utils/CharMeasure';
+import { CharMeasure } from './ui/CharMeasure';
 import { SelectionManager } from './SelectionManager';
 import { SelectionModel } from './SelectionModel';
 import { BufferSet } from './BufferSet';
