+++ src/ui/CharMeasure.test.ts
