+++ src/ui/CharMeasure.ts
